requirements.txt
Django==4.2.0
djangorestframework==3.14.0
requests==2.31.0
celery==5.3.6
python-dotenv==1.0.0
redis==4.5.5  # Required if using Redis as Celery brokerdrf-yasg==1.21.7
